requirements.txt
# Core
numpy>=1.24
torch>=2.0
transformers>=4.40
sentencepiece>=0.1.99

# Embeddings & Retrieval
faiss-cpu>=1.7.4
rank-bm25>=0.2.2
scikit-learn>=1.3

# NLP Utilities
nltk>=3.8

# Persistence & Utils
sqlite-utils>=3.35
tqdm>=4.66
pydantic>=2.5

# Optional (Serving / Future)
fastapi>=0.110
uvicorn>=0.27
